+++ CHANGELOG.md
@@ -7,6 +7,14 @@
 ### 🔨依赖升级
 
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.2.0)
+- spring-cloud-starter-alibaba-nacos-config(https://github.com/alibaba/spring-cloud-alibaba/tree/2.2.x/spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-nacos-config) 版本升级至 2.2.9.RELEASE
+- spring-cloud-starter-alibaba-nacos-discovery(https://github.com/alibaba/spring-cloud-alibaba/tree/2.2.x/spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-nacos-discovery) 版本升级至 2.2.9.RELEASE
+
+
+### ⭐ 新特性
+
+- **buession-springcloud-loadbalancer：** 新增 buession-springcloud-loadbalancer module
+- **buession-springcloud-nacos-discovery** 兼容处理 nacos discovery 在 undertow 容器中，停止服务时 java.lang.IllegalStateException: UT015023: This Context has been already 异常
 
 
 ---